GOAL-0005: Moved goals to private repository. Rework goals
