--- Content.Server/_Craft/StationGoals/Scripts/IStationGoalScript.cs
@@ -1,11 +1,0 @@
-using Robust.Shared.Prototypes;
-
-namespace Content.Server._Craft.StationGoals.Scipts;
-
-[ImplicitDataDefinitionForInheritors]
-public interface IStationGoalScript
-{
-    void PerformAction(StationGoalPrototype stationGoal, IPrototypeManager prototypeManager, EntitySystem entitySystem);
-
-    void Cleanup();
-}
