--- Content.Server/_Craft/StationGoals/StationGoalPaperComponent.cs
@@ -1,5 +1,0 @@
-namespace Content.Server._Craft.StationGoals;
-
-[RegisterComponent]
-public sealed class StationGoalPaperComponent : Component { }
-
