--- Content.Server/_Craft/Utils/ChatUtils.cs
@@ -1,40 +1,0 @@
-using Content.Server.Chat.Systems;
-
-namespace Content.Server._Craft.Utils;
-
-public static class ChatUtils
-{
-    public static void SendMessageFromCentcom(ChatSystem chatSystem, string message, EntityUid? stationId = null)
-    {
-        if (stationId == null)
-        {
-            chatSystem.DispatchGlobalAnnouncement(
-                message: message,
-                sender: "Центральное командование",
-                playSound: true,
-                colorOverride: Color.Yellow
-            );
-
-            return;
-        }
-
-        chatSystem.DispatchStationAnnouncement(
-            source: (EntityUid) stationId,
-            message: message,
-            sender: "Центральное командование",
-            playDefaultSound: true,
-            colorOverride: Color.Yellow
-        );
-    }
-
-    public static void SendLocMessageFromCentcom(ChatSystem chatSystem, string locCode, EntityUid? stationId = null)
-    {
-        var message = Loc.GetString(locCode);
-        if (message == null)
-        {
-            return;
-        }
-
-        SendMessageFromCentcom(chatSystem, (string) message, stationId);
-    }
-}
