--- Resources/Locale/ru-RU/_craft-station/station-goal-component.ftl
@@ -1,86 +1,0 @@
-station-goal-fax-paper-name = Цель станции
-station-goal-shuttle =
-    Цель вашей смены - постройка пилотируемого шаттла.
-    Требования к шатлу:
-    1) Комната пилота.
-    2) Медицинская комната.
-    3) Комната для отдыха персонала.
-    4) Комната охраны.
-    5) Автомономное энергоснабжение и атмос-система.
-    Обеспечьте шаттл всем необходимым.
-    Для выполнения цели в отдел снабжения добавлены дополнительные материалы для заказа.
-station-goal-security-shuttle =
-    Цель вашей смены - постройка пилотируемого шаттла для перевозки особо опасных заключенных.
-    Требования к шатлу:
-    1) Комната охраны.
-    2) 2 отдельные камеры.
-    3) Комната наблюдения за камерами.
-    4) Медицинская комната.
-    5) Автомономное энергоснабжение и атмос-система.
-    Обеспечьте шаттл всем необходимым.
-    Для выполнения цели в отдел снабжения добавлены дополнительные материалы для заказа.
-station-goal-science-shuttle =
-    Цель вашей смены - постройка пилотируемого шаттла для научных исследований.
-    Требования к шатлу:
-    1) Комната исследований артефактов.
-    2) Комната для исследований живых объектов.
-    3) Комната для отдыха персонала.
-    4) Автомономное энергоснабжение и атмос-система.
-    Обеспечьте шаттл всем необходимым.
-    Для выполнения цели в отдел снабжения добавлены дополнительные материалы для заказа.
-station-goal-medical-shuttle =
-    Цель вашей смены - постройка медицинского пилотируемого шаттла
-    Требования к шатлу:
-    1) Приемный покой для принятия пациентов.
-    2) Отдельная комната для тяжелобольных.
-    3) Отдел вирусологии.
-    4) Ординаторскую комнату.
-    5) Автомономное энергоснабжение и атмос-система.
-    Для выполнения цели в отдел снабжения добавлены дополнительные материалы для заказа.
-station-goal-zoo-shuttle =
-    Цель вашей смены - постройка пилотируемого шаттла для перевозки животных.
-    Требования к шатлу:
-    1) Как минимум 4 загона для различных типов животных.
-    2) Полноценная атмос-система.
-    3) 4 вида животных в загонах.
-    4) Продовольствие для животных.
-    5) Маленькая мед.команата для лечения животных.
-    6) Автомономное энергоснабжение и атмос-система.
-    Для выполнения цели в отдел снабжения добавлены дополнительные материалы для заказа.
-station-goal-solar-panels =
-    Цель вашей смены - организация системы запасного питания для станции на основе солнечных панелей.
-    Для этого вам понадобится заказать все необходимые материалы в отделе снабжения и после построить 2 новые ветки солнечных панелей инженерным отделом.
-    А так же обеспечить изолированность производимой ими энергии в новые 3 СМЭСа не подключенные к общей сети станции.
-station-goal-artifacts =
-    Цель вашей смены - исследование космических артефактов.
-    Артефакты для исследования прибудут на шаттле через 2 минуты.
-    Изученные артефакты необходимо доставить на шаттле эвакуации в специальных контейнерах с документацией внутри.
-    Рекомендуемые пункты документа с описанием артефакта:
-    - Наименование объекта
-    - Свойства объекта
-    - ФИ и должность ответственного
-    Документ должен быть удостоверен печатью научного руководителя.
-station-goal-artifacts-research-incomming =
-    Шаттл с артефактами пристыковался к станции.
-    Через 10 минут он улетит, оставшиеся на борту погибнут.
-station-goal-artifacts-research-remove =
-    Шаттл с артефактами улетает!
-    Последний шанс покинуть его!
-station-goal-bank =
-    Цель вашей смены - постройка орбитального хранилища с припасами и технологиями.
-    Хранилище должно быть размещено в космосе отдельно от основной станции, проследите за прочностью его конструкции, случайный метеорит не должен повредить его.
-    В хранилище необходимо разместить 4 ящика:
-    - ящик с продвинутыми медикаментами
-    - ящик с запасами лучших семян
-    - ящик-холодильник еды с высокой питательной ценностью
-    - ящик с ценными, но не уникальными платами
-    Проследите за сохранностью содержимого в хранилище до окончания смены.
-station-goal-zoo =
-    Цель вашей смены - постройка зоопарка.
-    Требования к зоопарку:
-    1) 4 вольера для различных видов животных.
-    2) Обустройство вольеров. (место для сна, принятия пищи, атмос-система, свет)
-    3) Вольеры должны быть с животными.
-station-goal-autonomous-security = Цель вашей смены - организация автономного электроснабжения отдела безопасности.
-station-goal-autonomous-medical = Цель вашей смены - организация автономного электроснабжения медицинского отдела.
-station-goal-autonomous-science = Цель вашей смены - организация автономного электроснабжения научного отдела.
